invite_tracker/invite-tracker-mtb/requirements.txt
discord
python-dateutil
psutil
parsedatetime
flask
asyncpg
psycopg2
asyncpg
mysql.connector
pytz
asyncio
